Move mempool implementation to separate module
