+++ hschain/HSChain/P2P/Internal.hs
@@ -44,6 +44,7 @@ import HSChain.Control.Class
 import HSChain.Control.Shepherd
 import HSChain.Control.Util
 import HSChain.Logger
+import HSChain.Mempool
 import HSChain.Monitoring
 import HSChain.Network.Types
 import HSChain.P2P.PeerState.Timer
