+++ hschain/HSChain/Run.hs
@@ -36,11 +36,11 @@ import HSChain.Blockchain.Internal.Engine
 import HSChain.Blockchain.Internal.Engine.Types
 import HSChain.Control.Class
 import HSChain.Logger
+import HSChain.Mempool
 import HSChain.Monitoring
 import HSChain.Network.Types
 import HSChain.P2P
 import HSChain.Store
-import HSChain.Store.STM
 import HSChain.Types
 import HSChain.Types.Merkle.Types
 import HSChain.Control.Delay
