+++ hschain/HSChain/Store.hs
@@ -63,13 +63,6 @@ module HSChain.Store (
   , mustRetrieveValidatorSet
   , retrieveSavedState
   , storeStateSnapshot
-    -- * Mempool
-  , MempoolCursor(..)
-  , Mempool(..)
-  , MempoolInfo(..)
-  , hoistMempoolCursor
-  , hoistMempool
-  , nullMempool
     -- * Blockchain state
   , BChStore(..)
     -- * Blockchain invariants checkers
@@ -77,8 +70,6 @@ module HSChain.Store (
   , checkStorage
   , checkProposedBlock
   ) where
-
-import qualified Katip
 
 import Codec.Serialise           (Serialise)
 import Control.Monad.Catch       (MonadMask,MonadThrow,MonadCatch)
@@ -97,10 +88,7 @@ import Data.Text                 (Text)
 import Data.Generics.Product.Fields (HasField'(..))
 import Data.Generics.Product.Typed  (HasType(..))
 import qualified Data.List.NonEmpty as NE
-import qualified Data.Aeson         as JSON
-import qualified Data.Aeson.TH      as JSON
 import Lens.Micro
-import GHC.Generics              (Generic)
 
 import HSChain.Types.Blockchain
 import HSChain.Types.Merkle.Types
@@ -165,96 +153,8 @@ initDatabase c = do
     Just () -> return ()
 
 
-----------------------------------------------------------------
--- Mempool
-----------------------------------------------------------------
-
--- | Statistics about mempool
-data MempoolInfo = MempoolInfo
-  { mempool'size      :: !Int
-  -- ^ Number of transactions currently in mempool
-  , mempool'added     :: !Int
-  -- ^ Number of transactions added to mempool since program start
-  , mempool'discarded :: !Int
-  -- ^ Number of transaction discarded immediately since program start
-  , mempool'filtered  :: !Int
-  -- ^ Number of transaction removed during filtering
-  }
-  deriving (Show,Generic)
-JSON.deriveJSON JSON.defaultOptions
-  { JSON.fieldLabelModifier = drop 8 } ''MempoolInfo
-
-instance Katip.ToObject MempoolInfo
-instance Katip.LogItem  MempoolInfo where
-  payloadKeys Katip.V0 _ = Katip.SomeKeys []
-  payloadKeys _        _ = Katip.AllKeys
-
--- | Cursor into mempool which is used for gossiping data
-data MempoolCursor m alg tx = MempoolCursor
-  { pushTransaction :: !(tx -> m (Maybe (Hashed alg tx)))
-    -- ^ Add transaction to the mempool. It's preliminary checked and
-    --   if check fails it immediately discarded. If transaction is
-    --   accepted its hash is computed and returned
-  , advanceCursor   :: !(m (Maybe tx))
-    -- ^ Take transaction from front and advance cursor. If cursor
-    -- points at the end of queue nothing happens.
-  }
-
--- | Mempool which is used for storing transactions before they're
---   added into blockchain. Transactions are stored in FIFO manner
-data Mempool m alg tx = Mempool
-  { peekNTransactions :: !(m [tx])
-    -- ^ Return transactions in mempool as lazy list. This operation
-    --   does not alter mempool state
-  , filterMempool     :: !(m ())
-    -- ^ Remove transactions that are no longer valid from mempool
-  , getMempoolCursor  :: !(m (MempoolCursor m alg tx))
-    -- ^ Get cursor pointing to be
-  , txInMempool       :: !(Hashed alg tx -> m Bool)
-    -- ^ Checks whether transaction is mempool
-  , mempoolStats      :: !(m MempoolInfo)
-    -- ^ Number of elements in mempool
-  , mempoolSize       :: !(m Int)
-    -- ^ Number of transactions in mempool
-  , mempoolSelfTest   :: !(m [String])
-    -- ^ Check mempool for internal consistency. Each returned string
-    --   is internal inconsistency
-  }
-
-hoistMempoolCursor :: (forall a. m a -> n a) -> MempoolCursor m alg tx -> MempoolCursor n alg tx
-hoistMempoolCursor fun MempoolCursor{..} = MempoolCursor
-  { pushTransaction = fun . pushTransaction
-  , advanceCursor   = fun advanceCursor
-  }
-
-hoistMempool :: Functor n => (forall a. m a -> n a) -> Mempool m alg tx -> Mempool n alg tx
-hoistMempool fun Mempool{..} = Mempool
-  { peekNTransactions = fun peekNTransactions
-  , filterMempool     = fun filterMempool
-  , getMempoolCursor  = hoistMempoolCursor fun <$> fun getMempoolCursor
-  , txInMempool       = fun . txInMempool
-  , mempoolStats      = fun mempoolStats
-  , mempoolSize       = fun mempoolSize
-  , mempoolSelfTest   = fun mempoolSelfTest
-  }
 
 
--- | Mempool which does nothing. It doesn't contain any transactions
---   and discard every transaction is pushed into it. Useful if one
---   doesn't need any mempool.
-nullMempool :: (Monad m) => Mempool m alg tx
-nullMempool = Mempool
-  { peekNTransactions = return []
-  , filterMempool     = return ()
-  , mempoolStats      = return $ MempoolInfo 0 0 0 0
-  , mempoolSize       = return 0
-  , txInMempool       = const (return False)
-  , getMempoolCursor  = return MempoolCursor
-      { pushTransaction = const $ return Nothing
-      , advanceCursor   = return Nothing
-      }
-  , mempoolSelfTest   = return []
-  }
 
 
 ----------------------------------------------------------------
