+++ hschain/HSChain/Store/STM.hs
@@ -5,176 +5,20 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 -- |
 module HSChain.Store.STM (
-    -- * Mempool
-    newMempool
     -- * Blockchain state storge
-  , newSTMBchStorage
+    newSTMBchStorage
   , snapshotState
   ) where
 
-import Control.Applicative
 import Control.Concurrent.STM
 import Control.Monad
 import Control.Monad.Catch       (MonadThrow)
 import Control.Monad.IO.Class
-import Control.Monad.Trans.Class
-import Control.Monad.Trans.Maybe
-import Data.Foldable
-import Data.List       (nub, sort)
-import Data.Maybe      (fromMaybe)
-import Data.Tuple      (swap)
-import Text.Printf
-import qualified Data.IntMap.Strict as IMap
-import           Data.IntMap.Strict   (IntMap)
-import qualified Data.Map.Strict    as Map
-import           Data.Map.Strict      (Map)
-import qualified Data.Set           as Set
-import           Data.Set             (Set)
 
-import HSChain.Crypto
 import HSChain.Store
 import HSChain.Types.Blockchain
 import HSChain.Types.Merkle.Types
 
-
-newMempool
-  :: forall m alg tx. (Show tx, Ord tx, Crypto alg, CryptoHashable tx, MonadIO m)
-  => (tx -> m Bool)
-  -> m (Mempool m alg tx)
-newMempool validation = do
-  varFIFO   :: TVar (IntMap tx)           <- liftIO $ newTVarIO IMap.empty
-  varRevMap :: TVar (Map tx Int)          <- liftIO $ newTVarIO Map.empty
-  varTxSet  :: TVar (Set (Hashed alg tx)) <- liftIO $ newTVarIO Set.empty
-  varMaxN   :: TVar Int                   <- liftIO $ newTVarIO 0
-  -- Counters for tracking number of transactions
-  varAdded     :: TVar Int <- liftIO $ newTVarIO 0
-  varDiscarded :: TVar Int <- liftIO $ newTVarIO 0
-  varFiltered  :: TVar Int <- liftIO $ newTVarIO 0
-  return Mempool
-    { peekNTransactions = toList <$> liftIO (readTVarIO varFIFO)
-    -- Filtering of transactions is tricky! Validation function is
-    -- monadic so we cannot call it inside atomically block. And when
-    -- we call it outside atomically block content of maps could change!
-    --
-    -- In order to overcome this problem we make following assumtions:
-    --
-    --  * All transactions added to mempool were at some point valid
-    --  * If transaction become invalid for whatever reason it stays
-    --    invalid forever
-    --
-    --  So basic algorithm is:
-    --   1) We take all transactions in mempool
-    --   2) Select invalid ones
-    --   3) Remove them in separate atomically block
-    --
-    --  This way any new transactions which were added during checking
-    --  are not affected.
-    , filterMempool = do
-        -- Invalid transactions
-        invalidTx <- filterM (\(_,tx) -> not <$> validation tx)
-                   . IMap.toList
-                 =<< liftIO (readTVarIO varFIFO)
-        -- Remove invalid transactions
-        liftIO $ atomically $ do
-          modifyTVar' varFIFO   $ \m0 ->
-            foldl' (\m (i,_) -> IMap.delete i m) m0 invalidTx
-          modifyTVar' varRevMap $ \m0 ->
-            foldl' (\m (_,tx) -> Map.delete tx m) m0 invalidTx
-          modifyTVar' varTxSet  $ \s0 ->
-            foldl' (\s(_,tx) -> Set.delete (hashed tx) s) s0 invalidTx
-          modifyTVar' varFiltered (+ length invalidTx)
-    --
-    , mempoolStats = liftIO $ atomically $ do
-        mempool'size      <- IMap.size <$> readTVar varFIFO
-        mempool'added     <- readTVar varAdded
-        mempool'discarded <- readTVar varDiscarded
-        mempool'filtered  <- readTVar varFiltered
-        return MempoolInfo{..}
-    --
-    , mempoolSize = do m <- liftIO $ readTVarIO varFIFO
-                       return $! IMap.size m
-    --
-    , txInMempool = \txHash -> liftIO $ do
-        txSet <- readTVarIO varTxSet
-        return $ txHash `Set.member` txSet
-    --
-    , getMempoolCursor = liftIO $ atomically $ do
-        varN <- newTVar 0
-        return MempoolCursor
-          { pushTransaction = \tx -> runMaybeT $ do
-              let discardSTM = do modifyTVar' varAdded     succ
-                                  modifyTVar' varDiscarded succ
-                  discard    = do liftIO $ atomically discardSTM
-                                  empty
-              -- Abort if tx is already in mempool
-              do revMap <- liftIO $ readTVarIO varRevMap
-                 when (tx `Map.member` revMap) discard
-              -- Validate TX
-              lift (validation tx) >>= \case
-                False -> discard
-                True  -> MaybeT $ liftIO $ atomically $ do
-                  -- Still it's possible that tx was added while we
-                  -- were checking its validity
-                  rmap <- readTVar varRevMap
-                  case tx `Map.notMember` rmap of
-                    False -> Nothing <$ discardSTM
-                    True  -> do
-                      let txHash = hashed tx
-                      modifyTVar' varAdded     succ
-                      n <- succ <$> readTVar varMaxN
-                      modifyTVar' varFIFO   $ IMap.insert n tx
-                      modifyTVar' varRevMap $ Map.insert tx n
-                      modifyTVar' varTxSet  $ Set.insert txHash
-                      writeTVar   varMaxN   $! n
-                      return $ Just txHash
-            --
-          , advanceCursor = liftIO $ atomically $ do
-              fifo <- readTVar varFIFO
-              n    <- readTVar varN
-              case n `IMap.lookupGT` fifo of
-                Nothing       -> return Nothing
-                Just (n', tx) -> do writeTVar varN $! n'
-                                    return (Just tx)
-          }
-    --
-    , mempoolSelfTest = liftIO $ atomically $ do
-        fifo <- readTVar varFIFO
-        rev  <- readTVar varRevMap
-        tx   <- readTVar varTxSet
-        maxN <- readTVar varMaxN
-        let nFIFO  = IMap.size fifo
-            nRev   = Map.size rev
-            nTX    = Set.size tx
-            lFifo  = IMap.toAscList fifo
-            lRev   = sort $ swap <$> Map.toList rev
-            maxKey = fst <$> IMap.lookupMax fifo
-        --
-        return $ concat
-          [ [ printf "Mismatch in rev.map. nFIFO=%i nRev=%i" nFIFO nRev
-            | nFIFO /= nRev
-            ]
-          , [ printf "Mismatch in tx.set. nFIFO=%i nTX=%i" nFIFO nTX
-            | nFIFO /= nTX
-            ]
-          , [ unlines [ "True TX / cached set"
-                      , show trueTX
-                      , show tx
-                      ]
-            | let trueTX = Set.fromList (hashed <$> toList fifo)
-            , trueTX /= tx
-            ]
-          , [ "Duplicate transactions present"
-            | let txs = toList fifo
-            , nub txs /= txs
-            ]
-          , [ printf "RevMap is not reverse of FIFO.\nFIFO: %s\nRevMap: %s" (show lFifo) (show lRev)
-            | lFifo /= lRev
-            ]
-          , [ printf "Max N less than max FIFO key. maxN=%i maxKey=%s" maxN (show maxKey)
-            | maxN < fromMaybe 0 maxKey
-            ]
-          ]
-    }
 
 ----------------------------------------------------------------
 -- Blockchain storage
